Move multicast test into udp test file
